refactor: split lexical and neural code into separate folders
