+++ scripts/filtering_experiment.py
